# GUS-5917 Optimize Simplify