--- a/src/common/evaluable.ts
+++ b/src/common/evaluable.ts
@@ -64,8 +64,8 @@ export interface Evaluable {
    */
   simplify(
     ctx: Context,
-    strictKeys?: string[],
-    optionalKeys?: string[]
+    strictKeys?: Set<string>,
+    optionalKeys?: Set<string>
   ): Result | Evaluable
 
   /**