--- a/src/index.ts
+++ b/src/index.ts
@@ -115,17 +115,19 @@ class Engine {
    *
    * @param {ExpressionInput} exp  Raw expression.
    * @param {Context} context Evaluation data context.
-   * @param {string[]} strictKeys keys to be considered present even if they are not present in the context
-   * @param {string[]} optionalKeys keys to be considered not present unless they are in the context or in
+   * @param {string[] | Set<string>} strictKeys keys to be considered present even if they are not present in the
+   *  context. Passing as a Set is recommended for performance reasons.
+   * @param {string[] | Set<string>} optionalKeys keys to be considered not present unless they are in the context or in
    *  `strictKeys`; when `strictKeys` is `undefined` and `optionalKeys` is an array, every key that is not in
-   *  `optionalKeys` is considered to be present and thus will be evaluated
+   *  `optionalKeys` is considered to be present and thus will be evaluated. Passing as a Set is recommended for
+   *  performance reasons.
    * @returns {Inpunt | boolean}
    */
   simplify(
     exp: ExpressionInput,
     context: Context,
-    strictKeys?: Set<string>,
-    optionalKeys?: Set<string>
+    strictKeys?: string[] | Set<string>,
+    optionalKeys?: string[] | Set<string>
   ): Input | boolean {
     const result = this.parse(exp).simplify(context, strictKeys, optionalKeys)
     if (isEvaluable(result)) {