--- a/src/operand/index.ts
+++ b/src/operand/index.ts
@@ -18,8 +18,8 @@ export abstract class Operand implements Evaluable {
    */
   abstract simplify(
     ctx: Context,
-    strictKeys?: string[],
-    optionalKeys?: string[]
+    strictKeys?: Set<string>,
+    optionalKeys?: Set<string>
   ): Result | Evaluable
 
   /**