--- a/src/operand/index.ts
+++ b/src/operand/index.ts
@@ -18,8 +18,8 @@ export abstract class Operand implements Evaluable {
    */
   abstract simplify(
     ctx: Context,
-    strictKeys?: Set<string>,
-    optionalKeys?: Set<string>
+    strictKeys?: string[] | Set<string>,
+    optionalKeys?: string[] | Set<string>
   ): Result | Evaluable
 
   /**