--- a/src/unsafe/simplify.ts
+++ b/src/unsafe/simplify.ts
@@ -1,30 +1,27 @@
-import {
-  isEvaluable,
-  OPERATOR_AND,
-  OPERATOR_DIVIDE,
-  OPERATOR_EQ,
-  OPERATOR_GE,
-  OPERATOR_GT,
-  OPERATOR_IN,
-  OPERATOR_LE,
-  OPERATOR_LT,
-  OPERATOR_MULTIPLY,
-  OPERATOR_NE,
-  OPERATOR_NOR,
-  OPERATOR_NOT,
-  OPERATOR_NOT_IN,
-  OPERATOR_OR,
-  OPERATOR_OVERLAP,
-  OPERATOR_PREFIX,
-  OPERATOR_PRESENT,
-  OPERATOR_SUBTRACT,
-  OPERATOR_SUFFIX,
-  OPERATOR_SUM,
-  OPERATOR_UNDEFINED,
-  OPERATOR_XOR,
-} from '..'
 import { Context, Evaluable } from '../common/evaluable'
-import { isString, isUndefined } from '../common/type-check'
+import { isEvaluable, isString, isUndefined } from '../common/type-check'
+import { OPERATOR as OPERATOR_DIVIDE } from '../expression/arithmetic/divide'
+import { OPERATOR as OPERATOR_MULTIPLY } from '../expression/arithmetic/multiply'
+import { OPERATOR as OPERATOR_SUBTRACT } from '../expression/arithmetic/subtract'
+import { OPERATOR as OPERATOR_SUM } from '../expression/arithmetic/sum'
+import { OPERATOR as OPERATOR_EQ } from '../expression/comparison/eq'
+import { OPERATOR as OPERATOR_GE } from '../expression/comparison/ge'
+import { OPERATOR as OPERATOR_GT } from '../expression/comparison/gt'
+import { OPERATOR as OPERATOR_IN } from '../expression/comparison/in'
+import { OPERATOR as OPERATOR_LE } from '../expression/comparison/le'
+import { OPERATOR as OPERATOR_LT } from '../expression/comparison/lt'
+import { OPERATOR as OPERATOR_NE } from '../expression/comparison/ne'
+import { OPERATOR as OPERATOR_NOT_IN } from '../expression/comparison/not-in'
+import { OPERATOR as OPERATOR_OVERLAP } from '../expression/comparison/overlap'
+import { OPERATOR as OPERATOR_PREFIX } from '../expression/comparison/prefix'
+import { OPERATOR as OPERATOR_PRESENT } from '../expression/comparison/present'
+import { OPERATOR as OPERATOR_SUFFIX } from '../expression/comparison/suffix'
+import { OPERATOR as OPERATOR_UNDEFINED } from '../expression/comparison/undefined'
+import { OPERATOR as OPERATOR_AND } from '../expression/logical/and'
+import { OPERATOR as OPERATOR_NOR } from '../expression/logical/nor'
+import { OPERATOR as OPERATOR_NOT } from '../expression/logical/not'
+import { OPERATOR as OPERATOR_OR } from '../expression/logical/or'
+import { OPERATOR as OPERATOR_XOR } from '../expression/logical/xor'
 import { Collection } from '../operand/collection'
 import { Reference } from '../operand/reference'
 import { Value } from '../operand/value'
@@ -53,8 +50,8 @@ import { simplifyXor } from './xor'
 export const unsafeSimplify = (
   context: Context,
   opts: Options,
-  strictKeys?: string[],
-  optionalKeys?: string[]
+  strictKeys?: Set<string>,
+  optionalKeys?: Set<string>
 ) => {
   const simplifyInput = (input: Input): Input | Evaluable => {
     // Value or Reference