# Description
- Optimize Simplify when checking for strict and optional keys
- With very large lists being added to simplification the performance tanks
- [x] Release a major version with the breaking change? Or investigate a performatic way to allow both?
  - Having the reference create the Set did not work
  - Also can't just deal with it in the illogical.simplify, since users can call Evaluable.simplify, which can start with any operand.
- ~~Unrelated, but also fixed a circular dependency warning that I missed from the build logs.~~ Merged in its own PR
- [x] Remove unsafeSimplify
- [x] Build
- [x] Docs

## QA
Using real test case from our API that has 2.9k optional keys and 1.3k strictKeys being provided:
### Before
```
illogical.parse in 6 ms
Evaluable.simplify in 66 ms
```
### After
```
illogical.parse in 6 ms
Evaluable.simplify in 4 ms
```

### Test Coverage
- [x] unit

#### References
- Closes GUS-5917