@atoledo 2025-07-17
The extraction of anonymous functions did seem to help
### Before
<img width="422" height="268" alt="image" src="https://github.com/user-attachments/assets/82f2d687-71d2-4a1f-af8e-29e120a815e8" />
### After
<img width="422" height="268" alt="image" src="https://github.com/user-attachments/assets/5080c20f-75a4-4bff-b7d4-38a9b58a8d2b" />