@rthreei 2025-07-17
> The extraction of anonymous functions did seem to help

The difference seems negligible.  Not worth touching the core logical code.