@atoledo 2025-07-17
> I would like to avoid this being a breaking change. Instead, could we have the function support both arrays and sets, and therefore be backwards compatible?

I tried, but this kills the performance since each Evaluable can be the starting point. They all need to check if the argument is stil an array. I can try again though